+++ src/game_sp/Display.h
@@ -13,12 +13,12 @@ extern int ScreenScrollXPos, ScreenScrollYPos;
 extern int ExplosionShake, ExplosionShakeMurphy;
 extern boolean NoDisplayFlag;
 
-extern long DisplayMinX, DisplayMaxX;
-extern long DisplayMinY, DisplayMaxY;
+extern int DisplayMinX, DisplayMaxX;
+extern int DisplayMinY, DisplayMaxY;
 
 extern void subDisplayLevel();
 extern void ScrollTo(int, int);
 extern void ScrollTowards(int, int);
-extern void SoftScrollTo(int, int, long, int);
+extern void SoftScrollTo(int, int, int, int);
 
 #endif /* DISPLAY_H */
